// src/api/api-type.ts
export type PairResponseType = {
    reserves: any;
    kLast: any;
    price0CumulativeLast: any;
    price1CumulativeLast: any;
    totalSupply: any;
    decimals: any;
    token0: any;
    token1: any;
};

export type TokenType = {
    symbol: any;
    name: any;
    balance: any;
    decimals: any;
};